App.vue: Add helper navbar and remove old routes from main nav bar
Things I do in this commit:
Make job-post pages have different container padding-tops since they have an additional nav bar which changes the structure.

Add a helper navbar that simply just checks if you're on a job post route. This is the lazy way, but it works. It has to be adjusted whenever routes change names though.

Missing things:
There is some weird on-focus CSS where the color is tied to the background-color of the buton. This part is annoying me so I'm leaving it be for now.